requirements.txt
dj-rest-auth==7.0.1
defusedxml==0.7.1
Django==5.2
django-allauth==0.53.0
django-cors-headers==4.0.0
django-debug-toolbar==5.1.0
django-filter==25.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.0
